Change float to decimal

+++ MercadoPagoSDK.Test/Core/MPIPNTest.cs
@@ -91,7 +91,7 @@ namespace MercadoPagoSDK.Test.Core
             Payer payer = new Payer();
             payer.Email = "mlovera@kinexo.com";
 
-            payment.TransactionAmount = (float)100;
+            payment.TransactionAmount = (decimal)100;
             payment.Token = GenerateSingleUseCardToken(); // 1 use card token
             payment.Description = "Pago de seguro";
             payment.PaymentMethodId = "visa";
